--- src/file-saved/issue.ts
@@ -1,7 +1,0 @@
-import * as vscode from "vscode"
-
-/** Issue represents an issue that Tikibase found */
-export interface Issue {
-  diagnostics: vscode.Diagnostic[]
-  file: vscode.Uri
-}
